Added an AutomaticPojoCodecProvider
Automatically creates a PojoCodec for any unseen class that contains
serializable / deserializable properties.

JAVA-2567
