+++ driver-async/src/examples/tour/PojoQuickTour.java
@@ -24,8 +24,6 @@ import com.mongodb.async.client.MongoCollection;
 import com.mongodb.async.client.MongoDatabase;
 import com.mongodb.client.result.DeleteResult;
 import com.mongodb.client.result.UpdateResult;
-import org.bson.codecs.configuration.CodecRegistry;
-import org.bson.codecs.pojo.PojoCodecProvider;
 
 import java.util.List;
 import java.util.concurrent.CountDownLatch;
@@ -36,8 +34,6 @@ import static com.mongodb.client.model.Filters.not;
 import static com.mongodb.client.model.Updates.combine;
 import static com.mongodb.client.model.Updates.set;
 import static java.util.Arrays.asList;
-import static org.bson.codecs.configuration.CodecRegistries.fromProviders;
-import static org.bson.codecs.configuration.CodecRegistries.fromRegistries;
 
 /**
  * The POJO QuickTour code example see: https://mongodb.github.io/mongo-java-driver/3.5/getting-started-pojo
@@ -59,14 +55,8 @@ public class PojoQuickTour {
             mongoClient = MongoClients.create(args[0]);
         }
 
-        // create codec registry for POJOs
-        CodecRegistry pojoCodecRegistry = fromRegistries(
-                fromProviders(PojoCodecProvider.builder().register(Person.class, Address.class).build()),
-                MongoClients.getDefaultCodecRegistry());
-
         // get handle to "mydb" database
-        MongoDatabase database = mongoClient.getDatabase("mydb").withCodecRegistry(pojoCodecRegistry);
-
+        MongoDatabase database = mongoClient.getDatabase("mydb");
 
         // get a handle to the "people" collection
         final MongoCollection<Person> collection = database.getCollection("people", Person.class);
