+++ driver-async/src/main/com/mongodb/async/client/MongoClients.java
@@ -37,6 +37,7 @@ import org.bson.codecs.IterableCodecProvider;
 import org.bson.codecs.MapCodecProvider;
 import org.bson.codecs.ValueCodecProvider;
 import org.bson.codecs.configuration.CodecRegistry;
+import org.bson.codecs.pojo.AutomaticPojoCodecProvider;
 
 import java.io.Closeable;
 
@@ -217,7 +218,8 @@ public final class MongoClients {
                     new IterableCodecProvider(new DocumentToDBRefTransformer()),
                     new MapCodecProvider(new DocumentToDBRefTransformer()),
                     new GeoJsonCodecProvider(),
-                    new GridFSFileCodecProvider()));
+                    new GridFSFileCodecProvider(),
+                    new AutomaticPojoCodecProvider()));
 
     private static StreamFactory getStreamFactory(final StreamFactoryFactory streamFactoryFactory,
                                                   final SocketSettings socketSettings, final SslSettings sslSettings,
