+++ driver/src/examples/tour/PojoQuickTour.java
@@ -23,8 +23,6 @@ import com.mongodb.client.MongoCollection;
 import com.mongodb.client.MongoDatabase;
 import com.mongodb.client.result.DeleteResult;
 import com.mongodb.client.result.UpdateResult;
-import org.bson.codecs.configuration.CodecRegistry;
-import org.bson.codecs.pojo.PojoCodecProvider;
 
 import java.util.List;
 
@@ -34,8 +32,6 @@ import static com.mongodb.client.model.Filters.not;
 import static com.mongodb.client.model.Updates.combine;
 import static com.mongodb.client.model.Updates.set;
 import static java.util.Arrays.asList;
-import static org.bson.codecs.configuration.CodecRegistries.fromProviders;
-import static org.bson.codecs.configuration.CodecRegistries.fromRegistries;
 
 /**
  * The POJO QuickTour code example see: https://mongodb.github.io/mongo-java-driver/3.5/getting-started-pojo
@@ -56,14 +52,8 @@ public class PojoQuickTour {
             mongoClient = new MongoClient(new MongoClientURI(args[0]));
         }
 
-        // create codec registry for POJOs
-        CodecRegistry pojoCodecRegistry = fromRegistries(
-                fromProviders(PojoCodecProvider.builder().register(Person.class, Address.class).build()),
-                MongoClient.getDefaultCodecRegistry());
-
         // get handle to "mydb" database
-        MongoDatabase database = mongoClient.getDatabase("mydb").withCodecRegistry(pojoCodecRegistry);
-
+        MongoDatabase database = mongoClient.getDatabase("mydb");
 
         // get a handle to the "people" collection
         MongoCollection<Person> collection = database.getCollection("people", Person.class);
